+++ StairFix/Properties/AssemblyInfo.cs
@@ -1,20 +1,21 @@
 ﻿using MelonLoader;
-using StairFix; // The namespace of your mod class
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 // ...
-[assembly: MelonInfo(typeof(TemplateClass), "StairFix", "1.0.0", "Baumritter")]
+[assembly: MelonInfo(typeof(StairFix.StairFixClass), StairFix.BuildInfo.ModName, StairFix.BuildInfo.ModVersion, StairFix.BuildInfo.Author)]
 [assembly: MelonGame("Buckethead Entertainment", "RUMBLE")]
+[assembly: VerifyLoaderVersion(0, 6, 2)]
+[assembly: MelonColor(200, 0, 200, 0)]
+[assembly: MelonAuthorColor(200, 0, 200, 0)]
 
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("StairFix")]
+[assembly: AssemblyTitle(StairFix.BuildInfo.ModName)]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("StairFix")]
+[assembly: AssemblyProduct(StairFix.BuildInfo.ModName)]
 [assembly: AssemblyCopyright("Copyright ©  2024")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -25,7 +26,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("bd6c4365-e862-4dc4-be25-edb2e35f4aae")]
+[assembly: Guid("caa17032-5000-4195-b343-c26492084360")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -37,5 +38,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+[assembly: AssemblyVersion(StairFix.BuildInfo.ModVersion)]
+[assembly: AssemblyFileVersion(StairFix.BuildInfo.ModVersion)]