+++ ThunderstoreStuff/CHANGELOG.md
@@ -1,2 +1,7 @@
+# Version 2.0.1
+- Package formatting fix
+# Version 2.0.0
+- Updated to MelonLoader 0.6.2
+- Changed melon name to "StairsFix"
 # Version 1.0.0
 - Created